Removed conversion

+++ internal/light-client/event-listener/listener.go
@@ -108,6 +108,6 @@ func (el *EventListener) ProcessLogs() {
 			continue
 		}
 
-		el.sampler.ProcessEvent(parsedURL.Host, uint64(event.BlockHeight))
+		el.sampler.ProcessEvent(parsedURL.Host, event.BlockHeight)
 	}
 }
